requirements.txt
apache-airflow-providers-google
pyarrow==6.0.1
pandas==1.3.5
requests==2.27.1